Abstract detection logic in seperate file
