+++ packages/cli/src/lib/bundler/bundle.ts
@@ -27,6 +27,7 @@ import { createConfig, resolveBaseUrl } from './config';
 import { BuildOptions } from './types';
 import { resolveBundlingPaths } from './paths';
 import chalk from 'chalk';
+import { writeDetectedPluginsModule } from './discover';
 
 // TODO(Rugvip): Limits from CRA, we might want to tweak these though.
 const WARN_AFTER_BUNDLE_GZIP_SIZE = 512 * 1024;
@@ -47,6 +48,12 @@ export async function buildBundle(options: BuildOptions) {
     isDev: false,
     baseUrl: resolveBaseUrl(options.frontendConfig),
     getFrontendAppConfigs: () => options.frontendAppConfigs,
+  });
+
+  await writeDetectedPluginsModule({
+    config: options.fullConfig,
+    entry: options.entry,
+    targetDir: options.targetDir,
   });
 
   const isCi = yn(process.env.CI, { default: false });
